backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
pydantic
pydantic-settings
playwright
httpx
python-dotenv
Pillow
supabase
daytona
anthropic
google-genai
